Urgent fixes to DMCCCA

+++ cca_zoo/deepmodels/deepwrapper.py
@@ -78,7 +78,7 @@ class DeepWrapper(_CCA_Base):
         num_params = sum(p.numel() for p in self.model.parameters())
         print('total parameters: ', num_params)
         best_model = copy.deepcopy(self.model.state_dict())
-        self.model.float().to(self.device)
+        self.model.double().to(self.device)
         min_val_loss = torch.tensor(np.inf)
         epochs_no_improve = 0
         early_stop = False
@@ -132,7 +132,7 @@ class DeepWrapper(_CCA_Base):
         self.model.train()
         train_loss = 0
         for batch_idx, (data, label) in enumerate(train_dataloader):
-            data = [d.float().to(self.device) for d in list(data)]
+            data = [d.to(self.device) for d in list(data)]
             loss = self._update_weights(*data)
             train_loss += loss.item()
         return train_loss / len(train_dataloader)
@@ -177,7 +177,7 @@ class DeepWrapper(_CCA_Base):
             param.grad = None
         total_val_loss = 0
         for batch_idx, (data, label) in enumerate(val_dataloader):
-            data = [d.float().to(self.device) for d in list(data)]
+            data = [d.to(self.device) for d in list(data)]
             loss = self.model.loss(*data)
             total_val_loss += loss.item()
         return total_val_loss / len(val_dataloader)
@@ -207,7 +207,7 @@ class DeepWrapper(_CCA_Base):
             test_dataloader = DataLoader(test_dataset, batch_size=len(test_dataset))
         with torch.no_grad():
             for batch_idx, (data, label) in enumerate(test_dataloader):
-                data = [d.float().to(self.device) for d in list(data)]
+                data = [d.to(self.device) for d in list(data)]
                 z = self.model(*data)
                 if batch_idx == 0:
                     z_list = [z_i.detach().cpu().numpy() for i, z_i in enumerate(z)]
@@ -222,7 +222,7 @@ class DeepWrapper(_CCA_Base):
         test_dataloader = DataLoader(test_dataset, batch_size=len(test_dataset))
         with torch.no_grad():
             for batch_idx, (data, label) in enumerate(test_dataloader):
-                data = [d.float().to(self.device) for d in list(data)]
+                data = [d.to(self.device) for d in list(data)]
                 x = self.model.recon(*data)
                 if batch_idx == 0:
                     x_list = [x_i.detach().cpu().numpy() for i, x_i in enumerate(x)]
